requirements.txt
Django==3.0.3
djangorestframework==3.14.0
drf_yasg==1.21.7
